api/requirements.txt
msal
azure-cosmos
Flask
gunicorn